requirements.txt
flask==2.3.2
flask-sqlalchemy==3.0.3
flask-httpauth==4.8.0
ldap3==2.9.1
python-dotenv==1.0.0
werkzeug==2.3.6